fix issue where rate-limit retry would cause http error
When a POST request is issued by the client, the request body is read
from the io.Reader. If the client was being rate-limited, on the second
attempt to try the request the io.Reader would be exhausted causing an
error to be returned by the http.Client. This fix creates a reader on
each request that can be used.